feat(satellite): expose guards to SDK
