+++ src/libs/satellite/src/sdk/core/mod.rs
@@ -1,5 +1,6 @@
 mod controllers;
 mod db;
+mod guards;
 mod hooks;
 mod ic;
 mod logs;
@@ -8,6 +9,7 @@ mod storage;
 
 pub use controllers::*;
 pub use db::*;
+pub use guards::*;
 pub use hooks::*;
 pub use ic::*;
 pub use logs::*;
